pyproject.toml
[project]
name = "my_devops_package"
version = "0.0.1"
authors = [{ name = "jphppd" }]
description = "Template project for Python"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "MIT"
license-files = ["LICENSE"]
dependencies = ["pyyaml>=6.0.1"]

[project.urls]
Homepage = "https://github.com/jphppd/devops-python"

[project.scripts]
my-devops-project = "my_devops_package.__main__:main"

[dependency-groups]
dev = ["pytest>=7.0.1", "pytest-cov>=4.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["my_devops_package", "my_other_package"]

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
addopts = "--showlocals --cov=."

[tool.coverage.report]
show_missing = true
